requirements.txt
pandas>=2.2.3
langchain>=0.3.7
langchain-core>=0.3.15
langchain-community>=0.3.5
langchain-anthropic>=0.2.4
langchain-ollama>=0.2.0
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3.2
rank_bm25>=0.2.2

PyYAML~=6.0.2